Channel ID is a u8

+++ src/adc.rs
@@ -195,8 +195,8 @@ impl ScanDir for ScanDown {
 macro_rules! adc_pin {
     ($PXi:ident, $i:expr) => {
         impl<RES, MODE> Channel<Adc<RES, MODE>> for $PXi<Analog> where MODE: RunMode, RES: Resolution {
-            type ID = u32;
-            fn channel() -> u32 { $i }
+            type ID = u8;
+            fn channel() -> u8 { $i }
         }
     };
 }
